requirements.txt
# Requirements to install.
celery==5.4.0
Django==5.1.5
django-celery-beat==2.7.0
django-celery-results==2.5.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
psycopg2-binary==2.9.10